requirements.txt
httpx
python-dotenv
typer[all]
rich
tenacity
